Add Super Linter to GitHub actions (#124)
* Add Super Linter to GitHub actions

* Fix README linter errors

* Lint entire code base

* Don't validate all codebase

* Fix lint errors in /ansible

* Fix sh linter errors

* Fix clang-formatter linter errors

* Fix clang-formatter linter errors

* Fix clang-formatter linter errors

* Fix action yaml format

* Validate all code

* Fix clang format errors

* Update action rules

* Update action rules

* Exclude paths from ansible lint

* Fix ansible linter errors

* Lint checks for python

* Fix python lint checks

* Fix python linter errors

* Lint checks for Google Java format

* Fix checkstyle linter errors

* Add TS and Docker linter checks

* Fix Dockerfile lint checks

* Have TS linter look for files correctly

* Fix tsconfig

* Don't lint using TS standard

* Fix ansible lint errors

+++ daemon/src/link/latency_target_data.h
@@ -1,14 +1,17 @@
+// Copyright (c) 2022 Stryke Force FRC 2767
 #pragma once
+
+#include <string>
 
 #include "upright_target_data.h"
 
 namespace deadeye {
 struct LatencyTargetData : public UprightTargetData {
   bool lights;
-  long long timestamp;
+  int64 timestamp;
 
   LatencyTargetData(std::string_view id, int sn, bool valid, cv::Rect bb,
-                    bool lights, long long timestamp);
+                    bool lights, int64 timestamp);
 
   [[nodiscard]] std::string Dump() const override;
 };
